fleetnode: simplify per ce-simplify findings
- Extract testutil.NewH2CServer / testutil.NewH2CClient so the five
  inline httptest.NewUnstartedServer(h2c.NewHandler(...)) + h2c
  transport copies across cmd/fleetnode and fleetnodebootstrap test
  files share one canonical form.
- Collapse the symmetric handshake.go WithTimeout pair into a
  withHandshakeTimeout helper; the previous inline pattern was easy
  to forget on a future third step.
- Swap the defer order in runControlSession so close(done) runs
  before the stream-close defer. The watcher goroutine now retires
  via its quiet path on every normal return, avoiding a redundant
  close race when ctx happens to cancel at the same moment.
- Replace the no-op `if err := writeLockOwnerPID(f); err != nil
  { _ = err }` block with `_ = writeLockOwnerPID(f)` plus a single
  one-liner explaining the best-effort intent.
- Trim narrative comments on newGatewayHTTPClient, runControlSession,
  reapOrphanedPlugins, and the stdout-logging Run shim. Keep only
  the non-obvious WHY.
